reorganizing codebase to put certain actions and components into unique folders
