--- BrightID-rn/ReactotronConfig.js
@@ -1,9 +1,0 @@
-import Reactotron from 'reactotron-react-native';
-import { reactotronRedux } from 'reactotron-redux';
-
-Reactotron.configure({ name: 'BrightID' }) // controls connection & communication settings
-  .useReactNative() // add all built-in react native plugins
-  .use(reactotronRedux())
-  .connect(); // let's connect!
-
-console.tron = Reactotron;
