+++ BrightID-rn/src/components/GroupsScreens/CurrentGroupAvatar.js
