+++ BrightID-rn/src/components/GroupsScreens/CurrentGroupCard.js
