+++ BrightID-rn/src/components/GroupsScreens/EligibleGroupAvatar.js
