+++ BrightID-rn/src/components/GroupsScreens/EligibleGroupCard.js
