+++ BrightID-rn/src/components/GroupsScreens/GroupsScreen.js
