+++ BrightID-rn/src/components/GroupsScreens/SearchGroups.js
