+++ BrightID-rn/src/components/OnboardingScreens/SignUp.js
