+++ BrightID-rn/src/components/OnboardingScreens/onboardingCards/MaintainPrivacy.js
