+++ BrightID-rn/src/components/WebrtcScreens/DisplayQRCodeScreen.js
