+++ BrightID-rn/src/components/WebrtcScreens/NearbyAvatar.js
