+++ BrightID-rn/src/components/WebrtcScreens/QRCodeScannerScreen.js
